Attacks now deal damage and lives are functional.
If a player is in the red area they will lose a life, and the lives are displayed on screen. If they move into a red area, they lose a life but will remain invincile while in the red. If they leave and them re-enter the red, they will lose invincibility and lose another life. Once the players lives hit 0, the player sprite will disapear but the game will not end. A game over state needs to be implemented next.